More test rationalization
